Renamed spec files
