+++ spec/gcp_helper.rb
