+++ spec/spec_helper.rb
@@ -9,7 +9,7 @@ require 'rspec'
 require 'vmpooler'
 require 'redis'
 require 'vmpooler/metrics'
-require 'clouddns_helper'
+require 'gcp_helper'
 
 def project_root_dir
   File.dirname(File.dirname(__FILE__))
